// src/hooks/api/report/enums.ts
export enum ReportType {
    DAILY_REPORT = 'DAILY_REPORT',
    WEEKLY_REPORT = 'WEEKLY_REPORT',
    MILESTONE_REPORT = 'MILESTONE_REPORT',
    DELIVERY_REPORT = 'DELIVERY_REPORT',
}

export enum ReportStatus {
    SUBMITTED = 'SUBMITTED',
    UNDER_REVIEW = 'UNDER_REVIEW',
    APPROVED = 'APPROVED',
    REJECTED = 'REJECTED',
    FINAL = 'FINAL',
    PENDING_ADMIN_CHECK = 'PENDING_ADMIN_CHECK',
    PENDING_COMPANY_REVIEW = 'PENDING_COMPANY_REVIEW',
}